synced up dated DWF
